# MySvelteApp.Client/api/schema/schemas.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export const AuthErrorResponseSchema = {
    type: 'object',
    properties: {
        message: {
            type: 'string'
        }
    }
} as const;

export const AuthSuccessResponseSchema = {
    type: 'object',
    properties: {
        token: {
            type: 'string'
        },
        userId: {
            type: 'integer'
        },
        username: {
            type: 'string'
        }
    }
} as const;

export const LoginRequestSchema = {
    type: 'object',
    properties: {
        password: {
            type: 'string'
        },
        username: {
            type: 'string'
        }
    }
} as const;

export const RandomPokemonResponseSchema = {
    type: 'object',
    properties: {
        image: {
            type: 'string'
        },
        name: {
            type: 'string'
        },
        type: {
            type: 'string'
        }
    }
} as const;

export const RegisterRequestSchema = {
    type: 'object',
    properties: {
        email: {
            type: 'string'
        },
        password: {
            type: 'string'
        },
        username: {
            type: 'string'
        }
    }
} as const;