# MySvelteApp.Client/api/schema/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type AuthErrorResponse = {
    message?: string;
};

export type AuthSuccessResponse = {
    token?: string;
    userId?: number;
    username?: string;
};

export type LoginRequest = {
    password?: string;
    username?: string;
};

export type RandomPokemonResponse = {
    image?: string;
    name?: string;
    type?: string;
};

export type RegisterRequest = {
    email?: string;
    password?: string;
    username?: string;
};

export type GetRandomPokemonData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/RandomPokemon';
};

export type GetRandomPokemonErrors = {
    /**
     * Internal Server Error
     */
    500: {
        [key: string]: string;
    };
};

export type GetRandomPokemonError = GetRandomPokemonErrors[keyof GetRandomPokemonErrors];

export type GetRandomPokemonResponses = {
    /**
     * OK
     */
    200: RandomPokemonResponse;
};

export type GetRandomPokemonResponse = GetRandomPokemonResponses[keyof GetRandomPokemonResponses];

export type PostAuthLoginData = {
    /**
     * Login Request
     */
    body: LoginRequest;
    path?: never;
    query?: never;
    url: '/auth/login';
};

export type PostAuthLoginErrors = {
    /**
     * Bad Request
     */
    400: AuthErrorResponse;
    /**
     * Unauthorized
     */
    401: AuthErrorResponse;
};

export type PostAuthLoginError = PostAuthLoginErrors[keyof PostAuthLoginErrors];

export type PostAuthLoginResponses = {
    /**
     * OK
     */
    200: AuthSuccessResponse;
};

export type PostAuthLoginResponse = PostAuthLoginResponses[keyof PostAuthLoginResponses];

export type PostAuthRegisterData = {
    /**
     * Register Request
     */
    body: RegisterRequest;
    path?: never;
    query?: never;
    url: '/auth/register';
};

export type PostAuthRegisterErrors = {
    /**
     * Bad Request
     */
    400: AuthErrorResponse;
    /**
     * Conflict
     */
    409: AuthErrorResponse;
};

export type PostAuthRegisterError = PostAuthRegisterErrors[keyof PostAuthRegisterErrors];

export type PostAuthRegisterResponses = {
    /**
     * OK
     */
    200: AuthSuccessResponse;
};

export type PostAuthRegisterResponse = PostAuthRegisterResponses[keyof PostAuthRegisterResponses];

export type ClientOptions = {
    baseUrl: string;
};